refactor(core): split workflow review repositories by use case
